Small improvements to improve RIPE Atlas fetching

+++ data/generate_csv.py
@@ -1,7 +1,11 @@
 import argparse
+import concurrent.futures
 import csv
 import datetime
+import itertools
 import os
+import shutil
+import sys
 
 
 from get_probes import get_countries, get_probes, write_probes
@@ -46,22 +50,38 @@ if __name__ == '__main__':
     time_start = datetime.datetime.fromisoformat(args.start)
     time_end = datetime.datetime.fromisoformat(args.end)
     output_directory = args.output_directory
-    os.makedirs(output_directory)
+    os.makedirs(output_directory, exist_ok=True)
 
     probes_filename = os.path.join(output_directory, 'probes.csv')
-    probes = get_probes(countries, ip_type)
-    print(probes)
-    with open(probes_filename, 'w') as probes_file:
-        write_probes(probes, ip_type, probes_file)
+    if os.path.exists(probes_filename):
+        with open(probes_filename, 'r') as probes_file:
+            probes = list(csv.DictReader(probes_file))
+    else:
+        probes = get_probes(countries, ip_type)
+        with open(probes_filename, 'w') as probes_file:
+            write_probes(probes, ip_type, probes_file)
 
     latency_directory = os.path.join(output_directory, 'latencies')
-    os.makedirs(latency_directory)
+    os.makedirs(latency_directory, exist_ok=True)
     time = time_start
+    times = []
+    times_next = []
+    latency_paths = []
     while time < time_end:
         time_string = ''.join(filter(str.isdigit, time.isoformat()))
         time_next = time + datetime.timedelta(hours=1)
-        write_measurements(
-            probes, os.path.join(latency_directory, time_string + '.csv'),
-            time, time_next
-        )
+
+        times.append(time)
+        times_next.append(time_next)
+        latency_paths.append(os.path.join(latency_directory, time_string + '.csv'))
+
         time = time_next
+
+    with concurrent.futures.ThreadPoolExecutor(max_workers=1) as executor:
+        executor.map(
+            write_measurements,
+            itertools.repeat(probes),
+            latency_paths,
+            times,
+            times_next
+        )
